Move template utility functions to local_template_manager.py
